The Grand UI Reorganization Update (GUIR)
Removed all references to an owning player from UI classes except in cases where it is explicitly needed (e.g. in MainMap for selection discrimination). This process reduced the Button class to doing nothing useful, so I deleted it and instead made all its former children directly subclass JButton. This saves a surprising number of getView() calls in the top level containers.

Note that this will have broken ScriptInterface. I am not pushing a fixed version of that in case my 'fix' breaks something else. All you *should* have to do is delete any references to an 'owner' player and replace any getView() calls of ScriptButtons with empty strings. I am pushing an updated ScriptButton put do have a look at it to make sure it's still what you expect it to be.