Prompts are Seemingly Alive (PSA)
To make a prompt, call the relevant method in Gooey.getPrompts(). If not having a reference to Gooey is a problem in a place where you want to generate a prompt, let me know.
There are gl hf prompts when you start a new game. I tried to implement the end turn prompt too but it broke actually ending turns so I will revisit it later.